cluster-ui: move db pages components to `/pages/databases/components`
Epic: none
Release note: None
