+++ pkg/ui/workspaces/cluster-ui/src/pages/databases/components/index.ts
@@ -1,0 +1,11 @@
+// Copyright 2025 The Cockroach Authors.
+//
+// Use of this software is governed by the CockroachDB Software License
+// included in the /LICENSE file.
+
+export * from "./nodeRegionsSelector/nodeRegionsSelector";
+export * from "./regionNodesLabel/regionNodesLabel";
+export * from "./tableMetadataLastUpdated/tableMetadataJobControl";
+export * from "./tableMetadataLastUpdated/tableMetadataJobProgress";
+export * from "./tooltipMessages";
+export * from "./grantsTable";
