+++ pkg/ui/workspaces/cluster-ui/src/pages/databases/components/regionNodesLabel/regionLabel.module.scss
