+++ pkg/ui/workspaces/cluster-ui/src/pages/databases/components/tableMetadataLastUpdated/tableMetadataJobControl.module.scss
