+++ pkg/ui/workspaces/cluster-ui/src/pages/databases/components/tableMetadataLastUpdated/tableMetadataJobProgress.module.scss
