+++ pkg/ui/workspaces/cluster-ui/src/pages/databases/index.ts
@@ -1,0 +1,8 @@
+// Copyright 2024 The Cockroach Authors.
+//
+// Use of this software is governed by the CockroachDB Software License
+// included in the /LICENSE file.
+
+export * from "./databasesPageRoot";
+export * from "./views/databaseDetails";
+export * from "./views/tableDetailsV2";
