+++ pkg/ui/workspaces/cluster-ui/src/pages/index.ts
@@ -1,1 +1,6 @@
+// Copyright 2025 The Cockroach Authors.
+//
+// Use of this software is governed by the CockroachDB Software License
+// included in the /LICENSE file.
+
 export * from "./databases";
